requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.0.8
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
captcha==0.7.1
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
cmake==3.25.2
cmdstanpy==1.2.5
colorama==0.4.5
contourpy==1.0.5
cryptography==45.0.5
curl_cffi==0.11.4
cycler==0.11.0
dacite==1.9.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==5.2
dlib==19.24.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.1.0
extra-streamlit-components==0.1.80
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.16.2
files==1.1.1
flatbuffers==25.2.10
fonttools==4.37.4
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gspread==6.2.1
h5py==3.13.0
holidays==0.76
htmlmin==0.1.12
httplib2==0.22.0
idna==3.6
ImageHash==4.3.1
imodels==2.0.0
importlib_resources==6.5.2
ipykernel==6.16.0
ipython==8.5.0
isodate==0.7.2
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.1
mistune==3.0.2
ml_dtypes==0.5.1
mlxtend==0.23.4
multimethod==1.12
multitasking==0.0.11
namex==0.0.8
narwhals==1.45.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.4.2
nltk==3.9.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.61.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.68
openpyxl==3.0.10
opt_einsum==3.4.0
optree==0.15.0
overrides==7.4.0
packaging==21.3
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peewee==3.18.1
phik==0.12.4
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==4.0.0
plotly==5.11.0
plotly-express==0.4.1
prometheus-client==0.19.0
prompt-toolkit==3.0.31
prophet==1.1.7
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.2
pure-eval==0.2.2
puremagic==1.28
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.13.0
PyJWT==2.10.1
pyparsing==3.0.9
pypi==2.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2025.2
PyWavelets==1.8.0
pywin32==304
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.31.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.13.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
sonopy==0.1.2
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.5.1
stanio==0.5.1
statsmodels==0.13.5
streamlit==1.46.1
streamlit-authenticator==0.2.2
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
terminado==0.18.0
tex==1.8
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.67.1
traitlets==5.14.0
tweepy==4.14.0
typeguard==4.4.2
types-python-dateutil==2.8.19.14
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.1.0
vader==0.0.3
vaderSentiment==3.3.2
visions==0.8.1
watchdog==6.0.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
xlrd==2.0.1
ydata-profiling==4.16.1
yfinance==0.2.65
